refactor: update action methods in CartController and related services - implement result pattern - improve validations

+++ Services/Cart/CartDeleterService.cs
@@ -1,14 +1,17 @@
 ﻿using RepositoryContracts;
 using ServiceContracts.Interfaces.ICart;
+using ServiceContracts.Results;
 
 namespace Services
 {
     public class CartDeleterService : ICartDeleterService
     {
         private readonly ICartRepository _cartRepository;
-        public CartDeleterService( ICartRepository cartRepository)
+        private readonly ICartGetterService _cartGetterService;
+        public CartDeleterService( ICartRepository cartRepository, ICartGetterService cartGetterService)
         {
             _cartRepository = cartRepository;
+            _cartGetterService = cartGetterService;
         }
 
         public async Task ClearCart(Guid userId)
@@ -16,9 +19,15 @@ namespace Services
             await _cartRepository.ClearCart(userId);
         }
 
-        public async Task<bool> RemoveFromCart(int productId, int cartId)
+        public async Task<CartItemResult> RemoveFromCart(int productId, Guid userId)
         {
-            return await _cartRepository.RemoveFromCart(productId, cartId);
+            if (productId <= 0)
+            {
+                return CartItemResult.Fail("Invalid product ID");
+            }
+            int cartId = await _cartGetterService.GetCartIdByUserId(userId);
+            await _cartRepository.RemoveFromCart(productId, cartId);
+            return CartItemResult.Ok("Product successfully deleted");
         }
     }
 }
